Added interface/ModuleIdFunctor.h and src/ModuleIdFunctor.cc
