+++ NTupleTools/src/NTupleStubs.cc
@@ -8,8 +8,6 @@
 #include "DataFormats/L1TrackTrigger/interface/TTTrack.h"
 #include "DataFormats/L1TrackTrigger/interface/TTTypes.h"
 #include "DataFormats/DetId/interface/DetId.h"
-#include "DataFormats/SiPixelDetId/interface/PXBDetId.h"
-#include "DataFormats/SiPixelDetId/interface/PXFDetId.h"
 #include "DataFormats/SiPixelDetId/interface/StackedTrackerDetId.h"
 #include "SimTracker/TrackTriggerAssociation/interface/TTClusterAssociationMap.h"
 #include "SimTracker/TrackTriggerAssociation/interface/TTStubAssociationMap.h"
@@ -39,28 +37,6 @@ unsigned findByRef(const edm::Handle<V>& handle, const edm::Ref<V,T>& ref) {
 }
 }
 
-// _____________________________________________________________________________
-uint32_t ModuleIdFunctor::operator() (const DetId& id) const {
-    uint32_t layer  = 999999;
-    uint32_t ladder = 999999;
-    uint32_t module = 999999;
-
-    if (id.subdetId() == (int) PixelSubdetector::PixelBarrel) {
-        PXBDetId pxbId(id);
-        layer  = pxbId.layer();
-        ladder = pxbId.ladder();
-        module = pxbId.module();
-
-    } else if (id.subdetId() == (int) PixelSubdetector::PixelEndcap) {
-        PXFDetId pxfId(id);
-        layer  = (pxfId.side() == 2) ? pxfId.disk() : pxfId.disk()+7;
-        ladder = pxfId.ring();
-        module = pxfId.module();
-    }
-
-    assert(layer != 999999 && ladder != 999999 && module != 999999);
-    return 10000*layer + 100*(ladder-1) + (module-1)/2;
-}
 
 // _____________________________________________________________________________
 NTupleStubs::NTupleStubs(const edm::ParameterSet& iConfig) :
